# optimize svg files