--- a/build.sh
+++ b/build.sh
@@ -1,0 +1,1 @@
+npx svgo -f logos/ -r --disable=convertPathData --multipass --pretty