--- a/logos/distros/Tumbleweed-green.svg
+++ b/logos/distros/Tumbleweed-green.svg
@@ -1,72 +1,8 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<!-- Created with Inkscape (http://www.inkscape.org/) -->
-
-<svg
-   xmlns:dc="http://purl.org/dc/elements/1.1/"
-   xmlns:cc="http://creativecommons.org/ns#"
-   xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-   xmlns:svg="http://www.w3.org/2000/svg"
-   xmlns="http://www.w3.org/2000/svg"
-   xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
-   xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape"
-   width="191.71371mm"
-   height="81.106239mm"
-   viewBox="0 0 679.30056 287.38432"
-   id="svg4143"
-   version="1.1"
-   inkscape:version="0.91 r13725"
-   sodipodi:docname="Tumbleweed-green.svg">
-  <defs
-     id="defs4145" />
-  <sodipodi:namedview
-     id="base"
-     pagecolor="#ffffff"
-     bordercolor="#666666"
-     borderopacity="1.0"
-     inkscape:pageopacity="0.0"
-     inkscape:pageshadow="2"
-     inkscape:zoom="0.35"
-     inkscape:cx="-136.77829"
-     inkscape:cy="186.54932"
-     inkscape:document-units="px"
-     inkscape:current-layer="layer1"
-     showgrid="false"
-     fit-margin-top="10"
-     fit-margin-left="10"
-     fit-margin-right="10"
-     fit-margin-bottom="10"
-     inkscape:window-width="1920"
-     inkscape:window-height="1016"
-     inkscape:window-x="1080"
-     inkscape:window-y="27"
-     inkscape:window-maximized="1" />
-  <metadata
-     id="metadata4148">
-    <rdf:RDF>
-      <cc:Work
-         rdf:about="">
-        <dc:format>image/svg+xml</dc:format>
-        <dc:type
-           rdf:resource="http://purl.org/dc/dcmitype/StillImage" />
-        <dc:title></dc:title>
-      </cc:Work>
-    </rdf:RDF>
-  </metadata>
-  <g
-     inkscape:label="Layer 1"
-     inkscape:groupmode="layer"
-     id="layer1"
-     transform="translate(-51.778294,-431.52721)">
-    <text
-       y="682.08691"
-       x="87.499268"
-       style="font-size:98.27079773px;font-family:SourceSansPro-Light;fill:#3ab6a9"
-       font-size="98.2708px"
-       id="text3">Tumbleweed</text>
-    <path
-       inkscape:connector-curvature="0"
-       style="fill:#3ab6a9"
-       d="m 401.76607,466.96028 c -9.4,0 -18.5,2.7 -26.6,7.3 -7.9,4.5 -14.8,10.6 -20.9,17.3 -2.4,2.6 -4.7,5.4 -7,8.2 -2.2,2.7 -4.2,5.5 -6.3,8.3 -1,1.4 -2.1,2.9 -3.1,4.3 -2.1,2.9 -4.2,5.7 -6.5,8.3 -2.2,2.4 -4.5,4.7 -6.8,7 -5.8,5.7 -12,11 -18.9,15.4 -5.4,3.4 -11.8,6.7 -18.4,6.7 -17.7,0 -33.5,-14 -33.5,-35.3 0,-21.3 15.2,-32.9 33.5,-32.9 10.6,0 19.9,4.5 29,12.3 l -4.9,5 11.9,3 11.9,3 -3.3,-11.8 -3.3,-11.8 -5.3,5.5 c -9.3,-8.3 -20.8,-15 -35.3,-15 -23.7,0 -43.2,17.1 -43.2,43.2 0,25 19.5,43.2 41.4,43.2 9.1,0 17.9,-2.7 25.6,-7.3 7,-4.1 13.1,-9.5 18.7,-15.4 1.9,-2 3.7,-4 5.5,-6 2.4,-2.6 4.6,-5.2 6.6,-8.2 0.9,-1.3 1.7,-2.6 2.6,-3.9 1.6,-2.4 3.1,-4.8 4.8,-7.1 3.2,-4.3 6.5,-8.5 10,-12.5 6,-6.7 12.8,-12.9 20.7,-17.1 6.3,-3.4 13.3,-5.2 20.5,-5.2 25,0 39,17.4 39,39.6 0,20.7 -13.4,37.8 -36.5,37.8 -13.2,0 -24.8,-3.6 -37,-14.1 l 5.7,-6.9 -12.1,-2 -12.1,-2 4.3,11.5 4.3,11.5 4.8,-5.8 c 11.8,10.7 24.7,17.7 41.5,17.7 28,0 46.3,-21.3 46.3,-48.1 -0.1,-28.2 -19,-47.7 -47.6,-47.7 z"
-       id="path5" />
-  </g>
+<svg xmlns="http://www.w3.org/2000/svg" width="724.587" height="306.543" viewBox="0 0 679.301 287.384">
+    <g fill="#3ab6a9">
+        <text y="682.087" x="87.499" font-size="98.271" font-family="SourceSansPro-Light" transform="translate(-51.778 -431.527)">
+            Tumbleweed
+        </text>
+        <path d="m 401.76607,466.96028 c -9.4,0 -18.5,2.7 -26.6,7.3 -7.9,4.5 -14.8,10.6 -20.9,17.3 -2.4,2.6 -4.7,5.4 -7,8.2 -2.2,2.7 -4.2,5.5 -6.3,8.3 -1,1.4 -2.1,2.9 -3.1,4.3 -2.1,2.9 -4.2,5.7 -6.5,8.3 -2.2,2.4 -4.5,4.7 -6.8,7 -5.8,5.7 -12,11 -18.9,15.4 -5.4,3.4 -11.8,6.7 -18.4,6.7 -17.7,0 -33.5,-14 -33.5,-35.3 0,-21.3 15.2,-32.9 33.5,-32.9 10.6,0 19.9,4.5 29,12.3 l -4.9,5 11.9,3 11.9,3 -3.3,-11.8 -3.3,-11.8 -5.3,5.5 c -9.3,-8.3 -20.8,-15 -35.3,-15 -23.7,0 -43.2,17.1 -43.2,43.2 0,25 19.5,43.2 41.4,43.2 9.1,0 17.9,-2.7 25.6,-7.3 7,-4.1 13.1,-9.5 18.7,-15.4 1.9,-2 3.7,-4 5.5,-6 2.4,-2.6 4.6,-5.2 6.6,-8.2 0.9,-1.3 1.7,-2.6 2.6,-3.9 1.6,-2.4 3.1,-4.8 4.8,-7.1 3.2,-4.3 6.5,-8.5 10,-12.5 6,-6.7 12.8,-12.9 20.7,-17.1 6.3,-3.4 13.3,-5.2 20.5,-5.2 25,0 39,17.4 39,39.6 0,20.7 -13.4,37.8 -36.5,37.8 -13.2,0 -24.8,-3.6 -37,-14.1 l 5.7,-6.9 -12.1,-2 -12.1,-2 4.3,11.5 4.3,11.5 4.8,-5.8 c 11.8,10.7 24.7,17.7 41.5,17.7 28,0 46.3,-21.3 46.3,-48.1 -0.1,-28.2 -19,-47.7 -47.6,-47.7 z" transform="translate(-51.778 -431.527)"/>
+    </g>
 </svg>