--- a/logos/teams/TEMPLATE-team.svg
+++ b/logos/teams/TEMPLATE-team.svg
@@ -1,56 +1,9 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<svg
-   xmlns:dc="http://purl.org/dc/elements/1.1/"
-   xmlns:cc="http://creativecommons.org/ns#"
-   xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-   xmlns:svg="http://www.w3.org/2000/svg"
-   xmlns="http://www.w3.org/2000/svg"
-   id="svg8"
-   version="1.1"
-   viewBox="0 0 180 82.000005"
-   height="82"
-   width="180">
-  <defs
-     id="defs2" />
-  <metadata
-     id="metadata5">
-    <rdf:RDF>
-      <cc:Work
-         rdf:about="">
-        <dc:format>image/svg+xml</dc:format>
-        <dc:type
-           rdf:resource="http://purl.org/dc/dcmitype/StillImage" />
-        <dc:title></dc:title>
-      </cc:Work>
-    </rdf:RDF>
-  </metadata>
-  <g
-     transform="translate(0,-1040.5197)"
-     id="layer1">
-    <path
-       id="rect1691"
-       transform="translate(0,1040.5197)"
-       d="M 0 0 L 0 82 L 180 82 L 180 0 L 0 0 z M 8 8 L 74 8 L 74 74 L 8 74 L 8 8 z "
-       style="opacity:1;fill:#35b9ab;fill-opacity:1;stroke:none;stroke-width:1.99718058;stroke-miterlimit:4;stroke-dasharray:none;stroke-opacity:1" />
-    <text
-       id="text1624"
-       y="1080.2151"
-       x="84"
-       style="font-style:normal;font-weight:normal;font-size:31.18781662px;line-height:1.25;font-family:sans-serif;letter-spacing:0px;word-spacing:0px;fill:#000000;fill-opacity:1;stroke:none;stroke-width:0.77969539"
-       xml:space="preserve"><tspan
-         style="font-style:normal;font-variant:normal;font-weight:normal;font-stretch:normal;font-family:'Open Sans Condensed';-inkscape-font-specification:'Open Sans Condensed, ';stroke-width:0.77969539"
-         y="1080.2151"
-         x="84"
-         id="tspan1622">Some</tspan></text>
-    <text
-       id="text1628"
-       y="1108.0905"
-       x="82.962746"
-       style="font-style:normal;font-weight:normal;font-size:35.76750946px;line-height:1.25;font-family:sans-serif;letter-spacing:0px;word-spacing:0px;fill:#000000;fill-opacity:1;stroke:none;stroke-width:0.89418781"
-       xml:space="preserve"><tspan
-         style="font-style:normal;font-variant:normal;font-weight:300;font-stretch:normal;font-family:'Source Sans Pro';-inkscape-font-specification:'Source Sans Pro Light';stroke-width:0.89418781"
-         y="1108.0905"
-         x="82.962746"
-         id="tspan1626">Team</tspan></text>
-  </g>
+<svg xmlns="http://www.w3.org/2000/svg" height="82" width="180">
+    <path d="M 0 0 L 0 82 L 180 82 L 180 0 L 0 0 z M 8 8 L 74 8 L 74 74 L 8 74 L 8 8 z" fill="#35b9ab"/>
+    <text y="1080.215" x="84" style="line-height:1.25" font-weight="400" font-size="31.188" font-family="sans-serif" letter-spacing="0" word-spacing="0" stroke-width=".78" transform="translate(0 -1040.52)">
+        <tspan style="-inkscape-font-specification:'Open Sans Condensed, '" y="1080.215" x="84" font-family="Open Sans Condensed">Some</tspan>
+    </text>
+    <text y="1108.091" x="82.963" style="line-height:1.25" font-weight="400" font-size="35.768" font-family="sans-serif" letter-spacing="0" word-spacing="0" stroke-width=".894" transform="translate(0 -1040.52)">
+        <tspan style="-inkscape-font-specification:'Source Sans Pro Light'" y="1108.091" x="82.963" font-weight="300" font-family="Source Sans Pro">Team</tspan>
+    </text>
 </svg>