… logos/buttons/button-colour.svg
… logos/buttons/button-filled-colour.svg
… logos/buttons/button-filled-monochrome.svg
… logos/buttons/button-laptop-colour.svg
… logos/buttons/button-laptop-monochrome.svg
… logos/buttons/button-monochrome.svg
… logos/distros/Evergreen-green.svg
… logos/distros/Leap-black-green.svg
… logos/distros/Leap-black.svg
… logos/distros/Leap-green.svg
… logos/distros/Leap-white-green.svg
… logos/distros/Leap-white.svg
… logos/distros/Tumbleweed-black-green.svg
… logos/distros/Tumbleweed-black.svg
… logos/distros/Tumbleweed-white-green.svg
… logos/events/hackweek-chameleon.svg
… logos/events/oSC19 logo with name.svg
… logos/official/geeko-color.svg
… logos/official/geeko-monochrome.svg
… logos/official/logo-color.svg
… logos/official/logo-monochrome.svg
… logos/teams/heroes-team.svg
… logos/teams/video-team.svg
… logos/unofficial/Design3.svg
… logos/unofficial/Design3_bw_outline.svg
… logos/unofficial/Design5.svg
… logos/unofficial/Design5_bw_outline.svg
… logos/unofficial/openSUSE-logos-suggestion.svg