```
button-colour.svg:
Done in 41 ms!
3.457 KiB - 36.4% = 2.198 KiB

button-filled-monochrome.svg:
Done in 12 ms!
3.773 KiB - 32.4% = 2.55 KiB

button-filled-colour.svg:
Done in 10 ms!
3.773 KiB - 32% = 2.564 KiB

button-monochrome.svg:
Done in 8 ms!
3.429 KiB - 35.7% = 2.203 KiB

button-laptop-colour.svg:
Done in 31 ms!
13.184 KiB - 9.3% = 11.953 KiB

button-laptop-monochrome.svg:
Done in 25 ms!
13.184 KiB - 9.5% = 11.936 KiB

Evergreen-green.svg:
Done in 53 ms!
23.058 KiB - 16.7% = 19.208 KiB

Leap-black.svg:
Done in 11 ms!
3.548 KiB - 66.5% = 1.189 KiB

Leap-black-green.svg:
Done in 10 ms!
3.681 KiB - 65.8% = 1.258 KiB

Leap-green.svg:
Done in 11 ms!
3.715 KiB - 65.4% = 1.286 KiB

Leap-white.svg:
Done in 10 ms!
3.949 KiB - 67.5% = 1.283 KiB

Tumbleweed-black-green.svg:
Done in 4 ms!
2.955 KiB - 57% = 1.27 KiB

Tumbleweed-black.svg:
Done in 10 ms!
2.942 KiB - 57.4% = 1.255 KiB

Leap-white-green.svg:
Done in 11 ms!
3.811 KiB - 66% = 1.294 KiB

Tumbleweed-green.svg:
Done in 7 ms!
2.979 KiB - 56.2% = 1.304 KiB

geeko-color.svg:
Done in 4 ms!
4.059 KiB - 1.8% = 3.987 KiB

geeko-monochrome.svg:
Done in 3 ms!
4.029 KiB - 2% = 3.948 KiB

oSC19 logo with name.svg:
Done in 182 ms!
34.242 KiB - 33% = 22.94 KiB

logo-color.svg:
Done in 10 ms!
7.752 KiB - 2.6% = 7.552 KiB

logo-monochrome.svg:
Done in 13 ms!
7.723 KiB - 2.8% = 7.509 KiB

artwork-team.svg:
Done in 29 ms!
6.565 KiB - 46.9% = 3.489 KiB

heroes-team.svg:
Done in 17 ms!
7.35 KiB - 30.7% = 5.093 KiB

video-team.svg:
Done in 24 ms!
17.614 KiB - 28.4% = 12.605 KiB

Tumbleweed-white.svg:
Done in 7 ms!
3.007 KiB - 56.8% = 1.299 KiB

Design3.svg:
Done in 12 ms!
2.654 KiB + 1.9% = 2.704 KiB

Design3_bw_outline.svg:
Done in 12 ms!
3.469 KiB - 14.3% = 2.972 KiB

Design5.svg:
Done in 15 ms!
4.139 KiB - 3.4% = 3.996 KiB

Design5_bw_outline.svg:
Done in 6 ms!
8.752 KiB + 0.6% = 8.805 KiB

openSUSE-logos-suggestion.svg:
Done in 308 ms!
106.507 KiB - 50.6% = 52.649 KiB

TEMPLATE-team.svg:
Done in 14 ms!
2.252 KiB - 59.9% = 0.902 KiB

Tumbleweed-white-green.svg:
Done in 8 ms!
2.997 KiB - 57.2% = 1.281 KiB

hackweek-chameleon.svg:
Done in 825 ms!
603.149 KiB - 16.7% = 502.651 KiB
```