@hellcp 2020-06-13
Remember that since git stores all of the file history, this doesn't improve cloning size in any way