@guoyunhe 2020-06-13
I know. This is only for download speed of CDN. (static.opensuse.org)