@hellcp 2020-06-13
We don't serve this repo in static either, and we probably shouldn't, since this is a development repo, and I cannot guarantee it will have the same structure tomorrow