@guoyunhe 2020-06-13
but currently we don't have one CDN for this purpose and it is needed. if this repo won't serve as CDN, I can create another one.